+++ Editor/Attributes/Utilities/DrawerUtilities.cs
@@ -1,7 +1,0 @@
-
-namespace BBUnity.Editor.Utilities {
-    class DrawerUtilities {
-
-       // public 
-    }
-}