+++ Runtime/BBMonoBehaviour.cs
@@ -4,22 +4,23 @@ using UnityEngine;
 namespace BBUnity {
 
     /// <summary>
-    /// The base behaviour for all components in BBUnity. This provides easy access methods for basic Unity
-    /// properties
+    /// An extended MonoBehaviour which provides access to a range of easier access / setters
+    /// and helper methods to aid the use of Coroutines. Please note that no virtual calls
+    /// to Awake, Update, OnGizmo are used in this class.
     /// </summary>
     [AddComponentMenu("")]
-    public class BaseBehaviour : MonoBehaviour {
+    public class BBMonoBehaviour : MonoBehaviour {
 
         /*
-         * Ease of use Properties
+         * Provides slightly easier ways to access the underlying MonoBehaviour
+         * propeties
          */
-        public string Name { get { return name; } set { Name = value; } }
         public Transform Parent { get { return transform.parent; } set { transform.parent = value; } }
-        public bool GameObjectActive { get { return gameObject.activeSelf; } set { gameObject.SetActive(value); } }
-        public bool GameObjectInactive { get { return !gameObject.activeSelf; } }
-        public bool Enabled { get { return enabled; } set { enabled = value; } }
-        public bool Disabled { get { return !enabled; } }
         public int Layer { get { return gameObject.layer; } set { gameObject.layer = value; } }
+        public bool IsGameObjectActive { get { return gameObject.activeSelf; } set { gameObject.SetActive(value); } }
+        public bool IsGameObjectInactive { get { return !gameObject.activeSelf; } }
+        public bool IsEnabled { get { return enabled; } set { enabled = value; } }
+        public bool IsDisabled { get { return !enabled; } }
 
         /*
          * Ease of use Setters
@@ -27,7 +28,7 @@ namespace BBUnity {
         public void SetName(string name) { this.name = name; }
         public void ActivateGameObject() { gameObject.SetActive(true); }
         public void DeactivateGameObject() { gameObject.SetActive(false); }
-        public void SetActiveGameObject(bool active) { gameObject.SetActive(active); }
+        public void SetGameObjectActive(bool active) { gameObject.SetActive(active); }
         public void Enable() { enabled = true; }
         public void Disable() { enabled = false; }
         public void SetEnabled(bool enable) { enabled = enable; }
@@ -35,7 +36,7 @@ namespace BBUnity {
         public void SetLayer(int layer) { gameObject.layer = layer; }
 
         /// <summary>
-        /// Resets the entire transformation with the following
+        /// Resets the entire transformation with the following:
         /// position = Vector3.zero
         /// rotation = Quaternion.identity;
         /// localScale = Vector3.one;
@@ -46,6 +47,10 @@ namespace BBUnity {
             transform.localScale = Vector3.one;
         }
 
+        /// <summary>
+        /// Resets the transformation with the option to skip the localScale
+        /// </summary>
+        /// <param name="resetScale"></param>
         public void ResetTransform(bool resetScale = true) {
             transform.position = Vector3.zero;
             transform.rotation = Quaternion.identity;
@@ -56,22 +61,37 @@ namespace BBUnity {
         }
 
         /// <summary>
-        /// Waits a given time in seconds and then calls the action provided
+        /// 
         /// </summary>
+        /// <param name="wait"></param>
+        /// <param name="action"></param>
         public void WaitThen(float wait, System.Action action) {
-            StartCoroutine(Wait(wait, action));
+            StartCoroutine(
+                Wait(wait, action)
+            );
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="wait"></param>
+        /// <param name="routine"></param>
         public void WaitThen(float wait, IEnumerator routine) {
             StartCoroutine(Wait(wait, () => {
                 StartCoroutine(routine);
             }));
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="wait"></param>
+        /// <param name="action"></param>
+        /// <returns></returns>
         private IEnumerator Wait(float wait, System.Action action) {
             yield return new WaitForSeconds(wait);
 
-            action();
+            action.Invoke();
         }
     }
 }