+++ Runtime/Entities/ICharacter.cs
@@ -1,0 +1,9 @@
+namespace BBUnity.Entities {
+
+    /// <summary>
+    /// Interface representing the base character
+    /// </summary>
+    public interface ICharacter : IEntity {
+
+    }
+}