+++ Runtime/Entities/IEntity.cs
@@ -1,0 +1,9 @@
+namespace BBUnity.Entities {
+
+    /// <summary>
+    /// Interface representing the base Entity provider
+    /// </summary>
+    public interface IEntity {
+
+    }
+}